feature(racing-game): CarRoster 를 별개의 파일로 분리
리뷰어님 피드백 반영.
