+++ src/main/kotlin/week1/step4/racinggame/CarRoster.kt
@@ -1,0 +1,14 @@
+package week1.step4.racinggame
+
+@JvmInline
+value class CarRoster(private val carNames: List<String>) {
+    init {
+        require(carNames.isNotEmpty()) {
+            "레이싱 게임을 하기 위해서는 차가 1대 이상이어야 합니다."
+        }
+    }
+
+    fun createCars(): List<Car> {
+        return carNames.map { Car(it) }
+    }
+}
